% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdots_validate_sims.R, R/feb_create_data.R
\name{addARerror}
\alias{addARerror}
\title{Create ar1 error}
\usage{
addARerror(val, rho = 0.8, sig = 0.025)

addARerror(val, rho = 0.8, sig = 0.025)
}
\description{
Create ar1 error

Create ar1 error
}
