% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{aggregateSub}
\alias{aggregateSub}
\title{Aggregates raw trial data into vwp style whatever}
\usage{
aggregateSub(x)
}
\arguments{
\item{x}{A subject run from simulation}
}
\value{
data.table with looks and proportions
}
\description{
Aggregates raw trial data into vwp style whatever
}
