% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{buildSaccadeSub}
\alias{buildSaccadeSub}
\title{Get start/end times from trial data for saccades}
\usage{
buildSaccadeSub(x)
}
\arguments{
\item{x}{A subject run from simulation}
}
\value{
data.table with entries being saccades and times
}
\description{
Get start/end times from trial data for saccades
}
