% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feb_create_data.R
\name{createData_feb}
\alias{createData_feb}
\title{Create artificial fixation data (February)}
\usage{
createData_feb(
  n = 25,
  trials = 100,
  pars = EMPIRICAL_START_PARS,
  paired = FALSE,
  pairMag = 0.05,
  ar1 = FALSE,
  manymeans = TRUE
)
}
\arguments{
\item{n}{number of subjects (it actually creates two groups of 25 each for bootstrap)}

\item{trials}{number of trials for binomial method}

\item{pars}{starting parameters for groups, gotten empirically. Probably will never change this value}

\item{paired}{is this paired data?}

\item{pairMag}{how much variability between paired subjects}

\item{ar1}{do i create ar1 data instead of binomial}

\item{manymeans}{do i do this the right way or the wrong way}
}
\description{
Create data for validating different situations in competing bdots implementations
ar = 0.8 and sd in ar noise is 0.025
}
\details{
Data to create "empirical" data centered around four parameter logistic
}
