% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataShiftLogistic.R
\name{createDataShiftLogistic}
\alias{createDataShiftLogistic}
\title{Create artificial fixation data}
\usage{
createDataShiftLogistic(
  n = 25,
  trials = 100,
  pars = EMPIRICAL_START_PARS,
  paired = TRUE,
  pairMag = 1,
  dd = 200
)
}
\arguments{
\item{n}{number of subjects (it actually creates two groups of 25 each for bootstrap)}

\item{trials}{number of trials for binomial method}

\item{pars}{starting parameters for groups, gotten empirically. Probably will never change this value}

\item{paired}{is this paired data?}

\item{pairMag}{how much variability between paired subjects}

\item{dd}{size of difference in crossover parameter}
}
\description{
Create data for validating different situations in competing bdots implementations
ar = 0.8 and sd in ar noise is 0.025
}
\details{
Data to create "empirical" data centered around four parameter logistic
}
