% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_power.R
\name{createPlineData2}
\alias{createPlineData2}
\title{Create piecewise linear data}
\usage{
createPlineData2(
  n = 25,
  trials = 100,
  ar1 = FALSE,
  pars = c(0, 0.5),
  manymeans = TRUE,
  TIME = seq(-0.5, 2, length.out = 501)
)
}
\arguments{
\item{n}{stuff}

\item{trials}{stuff}

\item{ar1}{stuff}

\item{pars}{stuff}

\item{manymeans}{stuff}

\item{TIME}{stuff}
}
\description{
Again, because maybe I'm retarded?
}
