% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{makeSubject}
\alias{makeSubject}
\title{Create single subject for VWP trial}
\usage{
makeSubject(fnct = "logistic")
}
\arguments{
\item{fnct}{character vector indicatin curve type}
}
\value{
This creates parameters for an individual subject,
returning the fixation curve parameters, as well as eye  movement
parameters for both standard and FBS+T
}
\description{
Create single subject for VWP trial
}
