% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset_fix_comparison_sim_functions.R
\name{makeSubject_compare}
\alias{makeSubject_compare}
\title{Create single subject for VWP trial}
\usage{
makeSubject_compare(fnct = "logistic", alt)
}
\arguments{
\item{fnct}{character vector indicatin curve type}

\item{alt}{TRUE/FALSE this determines if we use baseParams or baseParams2 for creating data}
}
\value{
This creates parameters for an individual subject,
returning the fixation curve parameters, as well as eye  movement
parameters for both standard and FBS+T
}
\description{
Create single subject for VWP trial
}
