% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{runSim}
\alias{runSim}
\title{Run simulation for multiple subjects}
\usage{
runSim(nsub = 10, ntrials = 300, fnct = "logistic", fbst = FALSE)
}
\arguments{
\item{nsub}{numeric, number of subjects in simulation}

\item{ntrials}{number of vwp trials per subject}

\item{fnct}{character vector indicatin curve type}

\item{fbst}{use FBS+T assumption or not}
}
\value{
This runs simluation for single subject, returns a list
containing information on subject, as well as trial data
}
\description{
Run simulation for multiple subjects
}
