% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_point_sim.R
\name{runSim_fixed}
\alias{runSim_fixed}
\title{Run simulation for multiple subjects}
\usage{
runSim_fixed(
  nsub = 10,
  ntrials = 300,
  fnct = "logistic",
  fbst = FALSE,
  sampDensity = 50,
  targMult = 1
)
}
\arguments{
\item{nsub}{numeric, number of subjects in simulation}

\item{ntrials}{number of vwp trials per subject}

\item{fnct}{character vector indicatin curve type}

\item{fbst}{use FBS+T assumption or not}

\item{sampDensity}{misnomer. Density for saccades}

\item{targMult}{Multiplier for target, i.e., multiple of sampDensity when fbst == TRUE}
}
\value{
This runs simluation for single subject, returns a list
containing information on subject, as well as trial data
}
\description{
Run simulation for multiple subjects
}
