% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{runSim_pb}
\alias{runSim_pb}
\title{Run simulation for multiple subjects}
\usage{
runSim_pb(
  nsub = 10,
  ntrials = 300,
  fnct = "logistic",
  fbst = FALSE,
  sacDelay = NULL
)
}
\arguments{
\item{nsub}{numeric, number of subjects in simulation}

\item{ntrials}{number of vwp trials per subject}

\item{fnct}{character vector indicatin curve type}

\item{fbst}{use FBS+T assumption or not}

\item{sacDelay}{alternative mechanism for eye delay}
}
\value{
This runs simluation for single subject, returns a list
containing information on subject, as well as trial data
}
\description{
Run simulation for multiple subjects
}
