% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_point_sim.R
\name{runSub_fixed}
\alias{runSub_fixed}
\title{Run simulation for single subject}
\usage{
runSub_fixed(
  fnct = "logistic",
  ntrials = 300,
  fbst = FALSE,
  sampDensity = 50,
  targMult = 1,
  window = NULL,
  windowRate = sampDensity,
  pars = NULL
)
}
\arguments{
\item{fnct}{character vector indicatin curve type}

\item{ntrials}{number of vwp trials per subject}

\item{fbst}{use FBS+T assumption or not}

\item{sampDensity}{misnomer. Density for saccades}

\item{targMult}{Multiplier for target, i.e., multiple of sampDensity when fbst == TRUE}

\item{window}{A time window to sample at different density}

\item{windowRate}{sampling density within window}

\item{pars}{curve parameters}
}
\value{
This runs simluation for single subject, returns a list
containing information on subject, as well as trial data
}
\description{
Run simulation for single subject
}
