% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootSimFunctions.R
\name{timeSliceFwer}
\alias{timeSliceFwer}
\title{Time Slice FWER rate}
\usage{
timeSliceFwer(y, f = mean)
}
\arguments{
\item{mm}{signifiance matrix}
}
\description{
Given list returned from getFWER (of all simulations) return function of
number of errors. For example, what is the mean/median number of of per comparison
error rate per simulation
}
