% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootSimFunctions.R
\name{timetie}
\alias{timetie}
\title{Signifiance matrix to boolean vector}
\usage{
timetie(mm)
}
\arguments{
\item{mm}{signifiance matrix}
}
\description{
Create length 401 boolean vector of where sig difference detected
}
