% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_power.R
\name{createPlineData}
\alias{createPlineData}
\title{Create piecewise linear data}
\usage{
createPlineData(
  n = 25,
  trials = 100,
  ar1 = FALSE,
  pars = c(0, 0.05),
  manymeans = TRUE,
  TIME = seq(-0.1, 1, length.out = 500),
  distSig = 0.025,
  paired = FALSE
)
}
\arguments{
\item{n}{stuff}

\item{trials}{stuff}

\item{ar1}{stuff}

\item{pars}{stuff}

\item{manymeans}{stuff}

\item{TIME}{stuff}
}
\description{
Again, because maybe I'm retarded? except this time wihtout "pairing" the base
and also yes i am retarded
}
