% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{runSub}
\alias{runSub}
\title{Run simulation for single subject}
\usage{
runSub(fnct = "logistic", ntrials = 300, fbst = FALSE)
}
\arguments{
\item{fnct}{character vector indicatin curve type}

\item{ntrials}{number of vwp trials per subject}

\item{fbst}{use FBS+T assumption or not}
}
\value{
This runs simluation for single subject, returns a list
containing information on subject, as well as trial data
}
\description{
Run simulation for single subject
}
